refactor: rename pipe-testing directory
