+++ packages/spectacular/src/lib/pipe-testing/test-pipe/test-pipe.component.ts
