+++ packages/spectacular/src/lib/pipe-testing/test-util/create-common-pipe-test-harness-options.ts
